Rename correctly stripNameAndValidate and stripName
