+++ src/Stub/AbstractStub.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 namespace Moka\Stub;
 
 use Moka\Exception\InvalidArgumentException;
-use function Moka\Stub\Helper\stripName;
+use function Moka\Stub\Helper\stripNameAndValidate;
 
 /**
  * Class AbstractStub
@@ -31,7 +31,7 @@ abstract class AbstractStub implements StubInterface
      */
     public function __construct(string $name, $value)
     {
-        $this->name = stripName($name);
+        $this->name = stripNameAndValidate($name);
         $this->value = $value;
     }
