+++ src/Stub/Helper/functions.php
@@ -2,13 +2,13 @@
 declare(strict_types=1);
 
 require_once __DIR__  . '/const.php';
-require_once __DIR__  . '/doStripName.php';
+require_once __DIR__ . '/stripName.php';
 require_once __DIR__  . '/isMethodName.php';
 require_once __DIR__  . '/isName.php';
 require_once __DIR__  . '/isName.php';
 require_once __DIR__  . '/isPropertyName.php';
 require_once __DIR__  . '/isStaticName.php';
-require_once __DIR__  . '/stripName.php';
+require_once __DIR__ . '/stripNameAndValidate.php';
 require_once __DIR__  . '/validateMethodName.php';
 require_once __DIR__  . '/validateName.php';
 require_once __DIR__  . '/validatePropertyName.php';
